refactor(aerorsync): move the config bridge to the application adapter
Fourth and last tranche of the AeroRsync standalone-crate plan, first of three commits. `from_rsync_config` was the module's remaining reason to name the application: it read an `RsyncConfig`, an AeroFTP type, and returned an `RsyncError`. It moves to the adapter as a free function, `aerorsync_adapter::config::transport_from_rsync_config`, with its body unchanged apart from the constructor it names.

A free function and not an inherent method on the transport, on purpose. Once the module becomes its own crate, an inherent block on a type that lives in another crate is not allowed, so an inherent method here would compile today and fail at extraction, which is the one thing this track is trying to avoid.

The six tests move with it. They read fields of the transport's SSH configuration, so the module exposes a read-only accessor, `AerorsyncDeltaTransport::ssh_config`, and the assertions are otherwise untouched: same profiles, same expectations, same names, now under `aerorsync_adapter::config::tests`.

One line changes outside the module: `providers/sftp.rs` calls the adapter function instead of the inherent constructor, and drops the import it no longer needs. `local_sync.rs`, `bin/aeroftp_cli.rs` and the integration tests keep compiling untouched, because the type they build and the trait they use are unchanged.

The module's application-import budget drops from 5 keys over 12 lines to 4 keys over 4 lines: the whole `rsync_over_ssh` entry for `delta_transport_impl.rs` is gone, imports included. What is left is the delta-transport trait in three test files and one `rsync_over_ssh` line in the local transport, which the next two commits take.

The adapter's own perimeter guard needed widening in the same commit, as its message demands: `config.rs` builds an `SshTransportConfig`, so `ssh_transport` joins the list of module paths the adapter may reach, with the reason written next to it.

The module suite goes from 688 to 689 passed, 0 failed, 15 ignored, zero skipped oracle pins, the difference being the adapter guard that A3 added and this tranche's plan did not count. `cargo fmt --all -- --check`, `cargo check --no-default-features --lib`, `cargo clippy --lib --features aerorsync --all-targets -- -D warnings` with and without the lane 3 flag, `cargo check --bin aeroftp-cli --features aerorsync`, the lane 3 compile check, `cargo test --lib delta_sync_rsync`, `cargo test --lib rsync_over_ssh`, `cargo test --lib local_sync`, the offline hard-rejection string contract and `git diff --check` all pass.
